+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - `cisco_fabricpath_topology` type and provider.
 - `cisco_itd_device_group` type and provider.
 - `cisco_itd_device_group_node` type and provider.
+- `cisco_itd_service` type and provider.
 - `cisco_stp_global` type and provider.
 
 #### NetDev Resources